+++ src/Kernel.php
@@ -9,9 +9,7 @@ declare(strict_types=1);
 
 namespace DecodeLabs\Clip;
 
-use DecodeLabs\Archetype;
 use DecodeLabs\Clip\Controller\Generic as GenericController;
-use DecodeLabs\Dictum;
 use DecodeLabs\Genesis\Context;
 use DecodeLabs\Genesis\Kernel as KernelInterface;
 use DecodeLabs\Terminus;
@@ -31,19 +29,8 @@ class Kernel implements KernelInterface
      */
     public function initialize(): void
     {
-        // Task name
-        Archetype::registerCustomNormalizer(
-            Task::class,
-            function (string $name): string {
-                $parts = explode('/', $name);
-
-                foreach ($parts as $i => $part) {
-                    $parts[$i] = Dictum::id($part);
-                }
-
-                return implode('\\', $parts);
-            }
-        );
+        // Normalizer
+        Normalizer::ensureRegistered();
 
         // Controller
         if (!$this->context->container->has(Controller::class)) {